refactor(commands): split driver_command.rs into focused submodules
