+++ src-tauri/src/commands/driver_command/tests.rs
@@ -1,807 +1,13 @@
-use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::Arc;
+use std::sync::Mutex;
 
-use super::error::{CmdExt, CommandError};
-use super::query::ensure_session_database;
-use super::AppState;
-use crate::mcp::permission::McpPermissionMode;
+use super::*;
+use super::helpers::inject_sql_target_fields;
 use datazen_driver_api::{
-    check_command_access, validate_command_input, CommandAccessLevel, CommandResult,
-    ConnectionHandle, DatabaseDriver, DriverCommandDefinition, DriverSaveDialogSpec,
-    QueryExecutionId, QueryStreamCallback, QueryStreamEvent,
+    CommandAccessLevel, QueryExecutionId, QueryStreamCallback, QueryStreamEvent,
 };
-use serde::{Deserialize, Serialize};
-use tauri::ipc::Channel;
-use tauri::State;
 
-pub(crate) fn access_level_for_mode(mode: Option<McpPermissionMode>) -> CommandAccessLevel {
-    match mode {
-        None | Some(McpPermissionMode::HighRiskWrite) => CommandAccessLevel::HighRisk,
-        Some(McpPermissionMode::SafeWrite) => CommandAccessLevel::Write,
-        Some(McpPermissionMode::ReadOnly) => CommandAccessLevel::Read,
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ExecuteDriverCommandStreamRequest {
-    #[serde(default)]
-    pub db_session_id: Option<String>,
-    pub command: String,
-    #[serde(default)]
-    pub input: serde_json::Value,
-    /// F1: optional explicit database pin — the session is switched to this
-    /// logical database before the command runs (same mechanism as the
-    /// query-family commands; `None`/blank keeps the current active database).
-    #[serde(default)]
-    pub database: Option<String>,
-    /// F7: optional target schema (PG-family engines). Rewrite-capable
-    /// drivers inline it as a qualified name; others ignore it.
-    #[serde(default)]
-    pub schema: Option<String>,
-    #[serde(default)]
-    pub apply_result_limit: Option<bool>,
-    #[serde(default)]
-    pub record_history: Option<bool>,
-}
-
-#[derive(Clone, Copy)]
-pub struct ExecuteDriverCommandStreamOpts {
-    pub apply_result_limit: bool,
-    pub record_history: bool,
-}
-
-impl Default for ExecuteDriverCommandStreamOpts {
-    fn default() -> Self {
-        Self {
-            apply_result_limit: true,
-            record_history: true,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ExecuteDriverCommandRequest {
-    #[serde(default)]
-    pub db_session_id: Option<String>,
-    #[serde(default)]
-    pub driver_type: Option<String>,
-    pub command: String,
-    #[serde(default)]
-    pub input: serde_json::Value,
-    /// F1: optional explicit database pin (session-bound commands only —
-    /// ignored for unbound `driverType` requests). See
-    /// `ensure_session_database` for the switching semantics.
-    #[serde(default)]
-    pub database: Option<String>,
-    /// F7: optional target schema (PG-family engines). Passed through into
-    /// the command input for SQL commands so rewrite-capable drivers can
-    /// inline it; ignored otherwise.
-    #[serde(default)]
-    pub schema: Option<String>,
-}
-
-fn unbound_handle() -> ConnectionHandle {
-    ConnectionHandle {
-        id: String::new(),
-        pool_id: String::new(),
-    }
-}
-
-fn nonempty(value: Option<&String>) -> Option<&str> {
-    value
-        .map(String::as_str)
-        .map(str::trim)
-        .filter(|s| !s.is_empty())
-}
-
-/// SQL SELECT row cap from settings. `None` when the "limit SELECT results"
-/// switch is off — streaming must not invent a cap from batch size.
-pub(crate) async fn query_result_limit_from_settings(state: &AppState) -> Option<u32> {
-    let settings = state.store.get_settings().await;
-    if settings.limit_select_results && settings.query_result_limit > 0 {
-        Some(settings.query_result_limit)
-    } else {
-        None
-    }
-}
-
-async fn apply_query_result_limit(state: &AppState, input: &mut serde_json::Value) {
-    if input.get("limit").is_some() {
-        return;
-    }
-    if let Some(limit) = query_result_limit_from_settings(state).await {
-        input["limit"] = serde_json::json!(limit);
-    }
-}
-
-fn sql_from_input(input: &serde_json::Value) -> Option<String> {
-    input
-        .get("sql")
-        .and_then(|v| v.as_str())
-        .filter(|s| !s.is_empty())
-        .map(str::to_string)
-}
-
-/// F7: copy non-blank envelope targeting fields into the command input object
-/// (no-op for blank values or non-object inputs).
-fn inject_sql_target_fields(
-    input: &mut serde_json::Value,
-    database: Option<&str>,
-    schema: Option<&str>,
-) {
-    let Some(map) = input.as_object_mut() else {
-        return;
-    };
-    let database = database.map(str::trim).filter(|s| !s.is_empty());
-    let schema = schema.map(str::trim).filter(|s| !s.is_empty());
-    if let Some(database) = database {
-        map.insert(
-            "database".into(),
-            serde_json::Value::String(database.into()),
-        );
-    }
-    if let Some(schema) = schema {
-        map.insert("schema".into(), serde_json::Value::String(schema.into()));
-    }
-}
-
-/// Persist a query/execute outcome to `{appData}/history.sqlite`.
-///
-/// The full SQL string is written **in plaintext** (see `history_db` module
-/// docs). Logging paths use `log_redact`; history storage does not yet.
-async fn record_sql_command_outcome(
-    state: &AppState,
-    db_session_id: Option<&str>,
-    sql: &str,
-    success: bool,
-    execution_time_ms: u64,
-    rows_affected: Option<u64>,
-    error_message: Option<String>,
-) {
-    let Some(db_session_id) = db_session_id else {
-        return;
-    };
-    // Resolve the persisted connection config that owns this runtime session.
-    let Some(connection_id) = state
-        .connection_manager
-        .owner_connection_id(db_session_id)
-        .await
-    else {
-        tracing::warn!(
-            db_session_id,
-            "Skipping history: no owning connectionId for this dbSessionId"
-        );
-        return;
-    };
-    // Record the session-active logical database so history can be grouped /
-    // filtered per panel context (empty string when the driver is single-db).
-    let database = state
-        .connection_manager
-        .get_session_config(db_session_id)
-        .await
-        .ok()
-        .and_then(|config| config.database)
-        .unwrap_or_default();
-    let entry = crate::store::QueryHistoryEntry {
-        id: uuid::Uuid::new_v4().to_string(),
-        connection_id,
-        database,
-        schema: None,
-        sql: sql.to_string(),
-        executed_at: chrono::Utc::now(),
-        execution_time_ms,
-        rows_affected,
-        success,
-        error_message,
-    };
-    // Persistence failures must not break query execution, but silently
-    // dropping history (disk full / locked db) is undebuggable — log it.
-    if let Err(e) = state.store.add_query_history(entry).await {
-        tracing::warn!(error = %e, "add_query_history failed");
-    }
-}
-
-fn query_rows_affected(data: &serde_json::Value) -> Option<u64> {
-    if let Some(rows) = data.get("rowsAffected").and_then(|v| v.as_u64()) {
-        return Some(rows);
-    }
-    let results = data.get("results")?.as_array()?;
-    Some(
-        results
-            .iter()
-            .filter_map(|r| r.get("rowsAffected").and_then(|v| v.as_u64()))
-            .sum(),
-    )
-}
-
-async fn resolve_command_driver(
-    state: &AppState,
-    db_session_id: Option<&String>,
-    driver_type: Option<&String>,
-) -> Result<(Arc<dyn DatabaseDriver>, ConnectionHandle, bool), CommandError> {
-    if let Some(id) = nonempty(db_session_id) {
-        let (driver, handle) = state
-            .connection_manager
-            .get_session(id)
-            .await
-            .cmd_err("execute_driver_command")?;
-        return Ok((driver, handle, true));
-    }
-    let driver_type = nonempty(driver_type)
-        .ok_or_else(|| CommandError::Validation("dbSessionId or driverType is required".into()))?;
-    let driver = state
-        .driver_registry
-        .get(&driver_type.to_string())
-        .await
-        .ok_or_else(|| CommandError::NotFound(format!("Driver not found: {driver_type}")))?;
-    Ok((driver, unbound_handle(), false))
-}
-
-pub(crate) async fn get_connection_commands_impl(
-    state: &AppState,
-    db_session_id: String,
-) -> Result<Vec<DriverCommandDefinition>, CommandError> {
-    let (driver, _) = state
-        .connection_manager
-        .get_session(&db_session_id)
-        .await
-        .cmd_err("get_connection_commands")?;
-    Ok(driver.command_definitions())
-}
-
-pub(crate) async fn get_driver_type_commands_impl(
-    state: &AppState,
-    driver_type: String,
-) -> Result<Vec<DriverCommandDefinition>, CommandError> {
-    let driver = state
-        .driver_registry
-        .get(&driver_type)
-        .await
-        .ok_or_else(|| CommandError::NotFound(format!("Driver not found: {driver_type}")))?;
-    Ok(driver.command_definitions())
-}
-
-pub(crate) async fn execute_driver_command_impl(
-    state: &AppState,
-    request: ExecuteDriverCommandRequest,
-) -> Result<CommandResult, CommandError> {
-    // Internal reuse (query/schema IPCs, MCP tools): never shows a native
-    // dialog — commands declaring a save dialog are rejected here.
-    execute_driver_command_with_mode(state, request, None, None).await
-}
-
-pub(crate) async fn execute_driver_command_stream_impl(
-    state: &AppState,
-    mut request: ExecuteDriverCommandStreamRequest,
-    on_event: QueryStreamCallback,
-    opts: ExecuteDriverCommandStreamOpts,
-) -> Result<(), CommandError> {
-    if request.command != "query_stream" {
-        return Err(CommandError::Validation(format!(
-            "Streaming is only supported for command 'query_stream', got '{}'",
-            request.command
-        )));
-    }
-
-    let db_session_id = request
-        .db_session_id
-        .as_ref()
-        .map(String::as_str)
-        .map(str::trim)
-        .filter(|s| !s.is_empty())
-        .ok_or_else(|| CommandError::Validation("dbSessionId is required".into()))?;
-
-    let (driver, handle, _bound) =
-        resolve_command_driver(state, request.db_session_id.as_ref(), None).await?;
-
-    // F1: pin the session's active database before streaming so unqualified
-    // SQL lands on the caller-selected database (no-op without a pin).
-    ensure_session_database(
-        state,
-        &handle.id,
-        request.database.as_deref(),
-        "execute_driver_command_stream",
-    )
-    .await?;
-
-    let definition = driver
-        .command_definitions()
-        .into_iter()
-        .find(|definition| definition.id == request.command)
-        .ok_or_else(|| {
-            CommandError::Validation(format!(
-                "Unsupported streaming driver command: {}",
-                request.command
-            ))
-        })?;
-
-    if opts.apply_result_limit {
-        apply_query_result_limit(state, &mut request.input).await;
-    }
-
-    validate_command_input(&definition, &request.input).map_err(CommandError::Validation)?;
-    check_command_access(&definition, CommandAccessLevel::Read)
-        .map_err(CommandError::Validation)?;
-
-    let sql = sql_from_input(&request.input).ok_or_else(|| {
-        CommandError::Validation("command 'query_stream' requires string input 'sql'".into())
-    })?;
-
-    if let Some(params) = request.input.get("params").cloned() {
-        let bound_sql =
-            crate::sql_guard::apply_params(&sql, &params).map_err(CommandError::Validation)?;
-        request.input["sql"] = serde_json::Value::String(bound_sql);
-    }
-    let mut sql = request
-        .input
-        .get("sql")
-        .and_then(|v| v.as_str())
-        .map(str::to_string)
-        .ok_or_else(|| {
-            CommandError::Validation("command 'query_stream' requires string input 'sql'".into())
-        })?;
-
-    // F7: give rewrite-capable drivers the chance to inline the SQL target
-    // (dialect-qualified names, no session switch). Drivers without the
-    // capability execute the SQL as-is and the ensure_session_database pin
-    // above remains the fallback for the database dimension.
-    let target_database = nonempty(request.database.as_ref()).map(str::to_string);
-    let target_schema = nonempty(request.schema.as_ref()).map(str::to_string);
-    if target_database.is_some() || target_schema.is_some() {
-        match driver.qualify_sql_target(&sql, target_database.as_deref(), target_schema.as_deref())
-        {
-            Some(qualified) => {
-                if qualified != sql {
-                    tracing::info!(
-                        db_session_id,
-                        "driver rewrote SQL target qualification for stream"
-                    );
-                }
-                sql = qualified;
-            }
-            None => {
-                tracing::debug!(
-                    db_session_id,
-                    database = target_database.as_deref().unwrap_or(""),
-                    schema = target_schema.as_deref().unwrap_or(""),
-                    "driver has no SQL target rewrite capability; streaming SQL as-is"
-                );
-            }
-        }
-    }
-
-    tracing::info!(
-        db_session_id,
-        sql_len = sql.len(),
-        "execute_driver_command_stream"
-    );
-    tracing::debug!(
-        db_session_id,
-        sql_preview = %crate::log_redact::sql_preview_for_log(&sql),
-        "execute_driver_command_stream sql"
-    );
-
-    let read_only = state
-        .connection_manager
-        .get_session_config(&handle.id)
-        .await
-        .map(|c| c.read_only)
-        .unwrap_or(false);
-    let safe_mode = state.store.get_settings().await.safe_mode;
-    crate::sql_guard::check_sql(&sql, read_only, safe_mode).map_err(CommandError::Validation)?;
-
-    // Every stream gets a fresh opaque execution identity. The driver's
-    // prepare hook registers a pending entry before the started event is
-    // published, so cancel can safely win the race before a backend target is
-    // acquired. Legacy drivers use the API's no-op/stream compatibility
-    // defaults, but never their old session-wide cancel method.
-    let execution_id = QueryExecutionId::new(uuid::Uuid::new_v4().to_string());
-    if let Err(error) = driver.prepare_query_execution(&handle, &execution_id).await {
-        let _ = driver.cleanup_query_execution(&handle, &execution_id).await;
-        return Err(error).cmd_err("execute_driver_command_stream");
-    }
-    if let Err(error) = state
-        .query_executions
-        .register(execution_id.clone(), handle.id.clone())
-        .await
-    {
-        let _ = driver.cleanup_query_execution(&handle, &execution_id).await;
-        return Err(CommandError::Validation(error));
-    }
-    on_event(QueryStreamEvent::ExecutionStarted {
-        execution_id: execution_id.as_str().to_string(),
-    });
-
-    let limit = if opts.apply_result_limit {
-        query_result_limit_from_settings(state).await
-    } else {
-        None
-    };
-    let rows_affected = Arc::new(AtomicU64::new(0));
-    let total_ms = Arc::new(AtomicU64::new(0));
-    let rows_cb = Arc::clone(&rows_affected);
-    let ms_cb = Arc::clone(&total_ms);
-    let user_cb = Arc::clone(&on_event);
-    let wrapped: QueryStreamCallback = Arc::new(move |event| {
-        match &event {
-            QueryStreamEvent::StatementEnd {
-                rows_affected: Some(n),
-                ..
-            } => {
-                rows_cb.fetch_add(*n, Ordering::Relaxed);
-            }
-            QueryStreamEvent::Done { total_time_ms } => {
-                ms_cb.store(*total_time_ms, Ordering::Relaxed);
-            }
-            _ => {}
-        }
-        user_cb(event);
-    });
-
-    let stream_result = driver
-        .query_stream_with_execution(&handle, &execution_id, &sql, limit, wrapped)
-        .await;
-    if let Err(error) = driver.cleanup_query_execution(&handle, &execution_id).await {
-        tracing::warn!(
-            db_session_id,
-            execution_id = %execution_id.as_str(),
-            error = %error,
-            "query execution driver cleanup failed"
-        );
-    }
-    state.query_executions.remove(&execution_id).await;
-
-    match stream_result {
-        Ok(()) => {
-            if opts.record_history {
-                record_sql_command_outcome(
-                    state,
-                    Some(db_session_id),
-                    &sql,
-                    true,
-                    total_ms.load(Ordering::Relaxed),
-                    Some(rows_affected.load(Ordering::Relaxed)),
-                    None,
-                )
-                .await;
-            }
-            if crate::cache::sql_may_mutate_schema(&sql) {
-                state.schema_cache.clear_connection(db_session_id).await;
-            }
-            Ok(())
-        }
-        Err(err) => {
-            if opts.record_history {
-                record_sql_command_outcome(
-                    state,
-                    Some(db_session_id),
-                    &sql,
-                    false,
-                    0,
-                    None,
-                    Some(err.to_string()),
-                )
-                .await;
-            }
-            Err(err).cmd_err("execute_driver_command_stream")
-        }
-    }
-}
-
-pub(crate) async fn execute_driver_command_with_mode(
-    state: &AppState,
-    mut request: ExecuteDriverCommandRequest,
-    permission_mode: Option<McpPermissionMode>,
-    dialog: Option<&tauri::AppHandle>,
-) -> Result<CommandResult, CommandError> {
-    let (driver, handle, bound) = resolve_command_driver(
-        state,
-        request.db_session_id.as_ref(),
-        request.driver_type.as_ref(),
-    )
-    .await?;
-
-    let definition = driver
-        .command_definitions()
-        .into_iter()
-        .find(|definition| definition.id == request.command)
-        .ok_or_else(|| {
-            CommandError::Validation(format!("Unsupported driver command: {}", request.command))
-        })?;
-
-    if !bound && definition.metadata.requires_connection {
-        return Err(CommandError::Validation(format!(
-            "Command '{}' requires a connection",
-            request.command
-        )));
-    }
-
-    // Save-dialog commands are interactive-only: the host thin shell pops the
-    // native dialog after execution, which is impossible headless (MCP /
-    // workflow / internal reuse). Reject before doing any work.
-    if definition.metadata.save_dialog.is_some() {
-        if permission_mode.is_some() || dialog.is_none() {
-            return Err(CommandError::Validation(format!(
-                "Command '{}' requires an interactive session to show its save dialog",
-                request.command
-            )));
-        }
-    }
-
-    // F1: session-bound commands honor an optional explicit database pin and
-    // switch the live session before execution; unbound driverType requests
-    // have no session to switch and ignore it.
-    if bound {
-        ensure_session_database(
-            state,
-            &handle.id,
-            request.database.as_deref(),
-            "execute_driver_command",
-        )
-        .await?;
-    }
-
-    let is_sql_command = matches!(definition.id.as_str(), "query" | "execute");
-    if definition.id == "query" {
-        apply_query_result_limit(state, &mut request.input).await;
-    }
-    // F7: pass envelope targeting into the command input so drivers that
-    // implement dialect-aware qualification (`qualify_sql_target`) can inline
-    // the target. Drivers without the capability ignore the extra fields and
-    // keep relying on the host session pin (ensure_session_database).
-    if is_sql_command {
-        inject_sql_target_fields(
-            &mut request.input,
-            request.database.as_deref(),
-            request.schema.as_deref(),
-        );
-    }
-
-    validate_command_input(&definition, &request.input).map_err(CommandError::Validation)?;
-    check_command_access(&definition, access_level_for_mode(permission_mode))
-        .map_err(CommandError::Validation)?;
-
-    if is_sql_command {
-        if let Some(params) = request.input.get("params").cloned() {
-            if let Some(sql) = request.input.get("sql").and_then(|v| v.as_str()) {
-                let bound_sql = crate::sql_guard::apply_params(sql, &params)
-                    .map_err(CommandError::Validation)?;
-                request.input["sql"] = serde_json::Value::String(bound_sql);
-            }
-        }
-        if bound {
-            if let Some(sql) = request.input.get("sql").and_then(|v| v.as_str()) {
-                let read_only = state
-                    .connection_manager
-                    .get_session_config(&handle.id)
-                    .await
-                    .map(|c| c.read_only)
-                    .unwrap_or(false);
-                let safe_mode = state.store.get_settings().await.safe_mode;
-                crate::sql_guard::check_sql(sql, read_only, safe_mode)
-                    .map_err(CommandError::Validation)?;
-            }
-        }
-        if let Some(mode) = permission_mode {
-            if let Some(sql) = request.input.get("sql").and_then(|v| v.as_str()) {
-                crate::mcp::permission::check_sql_allowed(sql, mode)
-                    .map_err(CommandError::Validation)?;
-            }
-        }
-    }
-
-    let sql = sql_from_input(&request.input);
-    let db_session_id = nonempty(request.db_session_id.as_ref()).map(str::to_string);
-    tracing::info!(
-        command = %request.command,
-        db_session_id = db_session_id.as_deref().unwrap_or(""),
-        sql_len = sql.as_ref().map(|s| s.len()).unwrap_or(0),
-        "execute_driver_command"
-    );
-    if let Some(sql) = sql.as_ref() {
-        tracing::debug!(
-            sql_preview = %crate::log_redact::sql_preview_for_log(sql),
-            "execute_driver_command sql"
-        );
-    }
-
-    match driver
-        .execute_command(&handle, &request.command, request.input)
-        .await
-    {
-        Ok(result) => {
-            if is_sql_command {
-                if let Some(sql) = sql.as_deref() {
-                    let execution_time_ms = result
-                        .data
-                        .get("totalTimeMs")
-                        .and_then(|v| v.as_u64())
-                        .unwrap_or(0);
-                    record_sql_command_outcome(
-                        state,
-                        db_session_id.as_deref(),
-                        sql,
-                        true,
-                        execution_time_ms,
-                        query_rows_affected(&result.data),
-                        None,
-                    )
-                    .await;
-                    if crate::cache::sql_may_mutate_schema(sql) {
-                        if let Some(id) = db_session_id.as_deref() {
-                            state.schema_cache.clear_connection(id).await;
-                        }
-                    }
-                }
-            }
-            // Generic host thin shell for metadata-declared save dialogs: no
-            // driver-type branching — any driver opting in gets the same flow.
-            if let Some(spec) = definition.metadata.save_dialog.as_ref() {
-                return finish_save_dialog(
-                    dialog.expect("save dialog checked above"),
-                    spec,
-                    result,
-                )
-                .await;
-            }
-            Ok(result)
-        }
-        Err(err) => {
-            if is_sql_command {
-                if let Some(sql) = sql.as_deref() {
-                    record_sql_command_outcome(
-                        state,
-                        db_session_id.as_deref(),
-                        sql,
-                        false,
-                        0,
-                        None,
-                        Some(err.to_string()),
-                    )
-                    .await;
-                }
-            }
-            Err(err).cmd_err("execute_driver_command")
-        }
-    }
-}
-
-/// Thin shell shared by every command whose metadata declares a save dialog:
-/// decode the command's byte payload, ask the user where to store it through a
-/// native save dialog, write the bytes and replace the result data with
-/// `{ <resultPathField>: savedPath | null }` (`null` on cancel). The absolute
-/// path is user-chosen in an OS dialog, mirroring the former dedicated
-/// `*_with_dialog` IPCs.
-async fn finish_save_dialog(
-    app: &tauri::AppHandle,
-    spec: &DriverSaveDialogSpec,
-    mut result: CommandResult,
-) -> Result<CommandResult, CommandError> {
-    use base64::{engine::general_purpose::STANDARD as BASE64, Engine};
-
-    if spec.extensions.is_empty() {
-        return Err(CommandError::Validation(
-            "Save-dialog command must declare at least one file extension".into(),
-        ));
-    }
-    let ext_list: Vec<&str> = spec.extensions.iter().map(String::as_str).collect();
-
-    let data_base64 = result
-        .data
-        .get(&spec.data_base64_field)
-        .and_then(serde_json::Value::as_str)
-        .ok_or_else(|| {
-            CommandError::Internal(format!(
-                "Save-dialog command did not return '{}' byte data",
-                spec.data_base64_field
-            ))
-        })?
-        .to_string();
-    let bytes = BASE64
-        .decode(data_base64.trim())
-        .map_err(|e| CommandError::Internal(format!("Command returned invalid base64: {e}")))?;
-    let file_name = result
-        .data
-        .get(&spec.file_name_field)
-        .and_then(serde_json::Value::as_str)
-        .filter(|name| !name.trim().is_empty())
-        .unwrap_or("download")
-        .to_string();
-
-    // Central dialog gateway: webdriver builds may replace this save dialog
-    // with an injected result (commands/dialog.rs).
-    let saved = match super::dialog::save_file(
-        app,
-        (spec.filter_name.clone(), spec.extensions.clone()),
-        file_name,
-    )
-    .await?
-    {
-        None => None,
-        Some(path) => {
-            super::file::validate_extension(&path, &ext_list)?;
-            let byte_count = bytes.len();
-            tokio::fs::write(&path, bytes)
-                .await
-                .cmd_err("execute_driver_command")?;
-            tracing::info!(
-                bytes = byte_count,
-                saved = %path.display(),
-                "driver command payload saved via native dialog"
-            );
-            Some(path.to_string_lossy().into_owned())
-        }
-    };
-
-    let mut data = serde_json::Map::new();
-    data.insert(
-        spec.result_path_field.clone(),
-        saved.map(serde_json::Value::String).unwrap_or_default(),
-    );
-    result.data = serde_json::Value::Object(data);
-    Ok(result)
-}
-
-/// Discover commands from a Driver type. No live Connection is required.
-#[tauri::command]
-pub async fn get_driver_commands(
-    state: State<'_, AppState>,
-    driver_type: String,
-) -> Result<Vec<DriverCommandDefinition>, CommandError> {
-    get_driver_type_commands_impl(&state, driver_type).await
-}
-
-/// Discover commands from a concrete live DB session.
-#[tauri::command]
-pub async fn get_connection_commands(
-    state: State<'_, AppState>,
-    db_session_id: String,
-) -> Result<Vec<DriverCommandDefinition>, CommandError> {
-    get_connection_commands_impl(&state, db_session_id).await
-}
-
-#[tauri::command]
-pub async fn execute_driver_command(
-    app: tauri::AppHandle,
-    state: State<'_, AppState>,
-    request: ExecuteDriverCommandRequest,
-) -> Result<CommandResult, CommandError> {
-    // GUI IPC passes the AppHandle so metadata-declared save dialogs can run.
-    execute_driver_command_with_mode(&state, request, None, Some(&app)).await
-}
-
-#[tauri::command]
-pub async fn execute_driver_command_stream(
-    state: State<'_, AppState>,
-    request: ExecuteDriverCommandStreamRequest,
-    on_event: Channel<QueryStreamEvent>,
-    apply_result_limit: Option<bool>,
-    record_history: Option<bool>,
-) -> Result<(), CommandError> {
-    let callback: QueryStreamCallback = Arc::new(move |event| {
-        let _ = on_event.send(event);
-    });
-    execute_driver_command_stream_impl(
-        &state,
-        request,
-        callback,
-        ExecuteDriverCommandStreamOpts {
-            apply_result_limit: apply_result_limit.unwrap_or(true),
-            record_history: record_history.unwrap_or(true),
-        },
-    )
-    .await
-}
-
-#[cfg(test)]
-mod tests {
-    use std::sync::Mutex;
-
-    use super::*;
-
-    #[test]
+#[test]
     fn request_uses_camel_case_wire_format() {
         let request = ExecuteDriverCommandRequest {
             db_session_id: Some("mysql-prod".into()),
@@ -1570,4 +776,3 @@ mod tests {
         .unwrap();
         assert!(result.data.is_object());
     }
-}
